Add create item functionality to items picker.
When the user clicks "Add Item", they are faced with a popup form
prompting them to enter a name for the item and to choose its
category from a drop down. Upon submitting this form, the new item
is persisted to the database and, if successful, the items picker
is updated.