Add create bin functionality.
This requires reworking the way we represent bins in the browser.
The old serve-rendered template is no more, we now dynamically
render the bins from a JSON representation that we update as needed.
This is similar to how we represented items/categories.

Note that this patch does not allow bins to be created for new cities,
only for existing cities being edited. This is because the city to which
a bin belongs must be known at bin creation time. This is not a problem,
because we plan to disable the new city function anyways.