Fix "unrecognized expression" JQuery error.
The init_smooth_scrolling code was operating on the selector 'nav a',
which included both the global navbar and the city-specific navbar.
It made some assumptions about the DOM which hold only for the city
specific one, causing errors when operating on the global one.

This patch adds a unique ID to the city-specific navbar, so the initialization
code can act on only the elements it needs to.