Add color to bins.
The Bin model contains a property "color", which is a hexadecimal
string of the form #ffffff. The city editor has been altered to
allow users to choose and edit the color of their bins.

Making this change to the GUI took some significant refactoring, as
our original design -- "every time something happens, clobber everything
and redraw it" -- was not amenable to extension. The major differences are:
 * No DOM elements are destroyed, except when deleting bins. When switching
   between the bin form and bin display, one is hidden and the other is shown,
   but both still exist in the DOM.
 * Only updated bins are refreshed. When something happens to one bin, we
   don't touch the other bins.
 * The structure, style, and behaviors of the bins are separated, as they
   should be, into HTML, CSS, and CoffeeScript. The DOM elements to
   display a bin are no longer created dynamically in JQuery. Instead, we
   have a generic "skeleton bin" in bins/_form.html/erb. This skeleton is
   rendered as a hidden element at the end of the body of the page.
   Whenever we need a new bin, we simply clone this structure and add only
   bin-specific dynamic properties.