Add basic information to a landing screen for cities.
The landing page displays at the top of the show city page, above
the bins. It includes the name, state, and zip code for the city,
as well as a link to the city's website and contact information,
if available.

The search bar is also moved here from the landing bar.